Enterキーの操作を管理するフックを追加し、関連するコンポーネントのロジックを整理。ResultLineコンポーネントのキーをインデックスに変更。ポップアップの動作を非同期に修正。
